Refactor Epoll structure/class Epoll class is in the separate file now. Added RAII mechanism
